Revert "optimize dive.explore a bit" (more)
This reverts commit 49d59f563bbeca9c5c180cb60301e741a940e99a.

Wow that last commit upped the time from 6s to 21s in release mode.
(to generate 1000 seeds)

I suppose this means that each `explore` doesn't usually involve many
passes, so the inefficiency didn't matter. Or maybe rust is doing smart
with immutable vecs that it can't do with mutable vecs. or something

edit: yep, it looks like it's normally 1~3 passes per explore

HOWEVER that will change once there are 25 items instead of 3 items,
so maybe re-revert this later and it'll actually help?